- [ZBX-3747] Fixed additional compiler warnings
git-svn-id: svn://svn.zabbix.com/branches/dev@19344 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-3747/src/libs/zbxsysinfo/sysinfo.c
@@ -230,7 +230,7 @@ void	free_result(AGENT_RESULT *result)
  *               1 - command without parameters;
  *               2 - command with parameters
  */
-int	parse_command(const char *command, char *cmd, int cmd_max_len,
+int	parse_command(const char *command, char *cmd, size_t cmd_max_len,
 		char *param, int param_max_len)
 {
 	char	*pl, *pr;
